chore(style): add script to rustfmt `quote!`
Script `rustfmt-quote.py` is stupid but it works.

I have tried to write it in Bash, but it is too slow. The reason to choose
Python is that it is built-in in most system.

The script first extracts the code between `quote!` and replace the macro variables
in it.

Then it surrounds the code in the correct context, and run `rustfmt` against
it.

However, some macro code still fails to pass `rustfmt`, so a blacklist is used to skip these kinds of code.